rename `images\service.py` file into `thumbnail.py`
