+++ src/sts/images/thumbnail.py
